refactor: have dev server depend on app-node
This flips the dependency between the dev-server and app-node. Instead of the having app-node depend on the dev-server as a library, the dev-server is now a standalone app which imports app-node.

This is in preparation of having the dev-server manage the entire dev environment including the wallet UI vite server and eventually app-beacon as well.