feat(oer): octet string with contained type
This allows the user to specific a type that is contained inside of a given octet string. When parsing, the contained type will be parsed as well. During serialization, either the value for the contained type can be passed or a Uint8Array representing the already-serialized value.

In order to support this, OerType now has separate types for the value that parsing returns and the value that you can give to serialize.