feat(dev): connect frontend to runners directly
The previous way of forwarding all events globally to the dev server, caching them there, and then subscribing to this global firehose and filtering was extremely inefficient. Nodes had to cache all events in case they were requested later, which creates a memory leak.

The new way is that the dev frontend simply connects directly to each runner and requests the needed information.