refactor(reactive): sig.use => sig.run
Previously, fetching a value from the reactor was "reactor.useContext" and instantiating an effect was "sig.use". Now that we have a helper for retrieving and automatically returning context references, we would like to call it "sig.use".

So the helper for instantiating effects is being renamed to "sig.run", which may also be more accurate in describing what it does.