refactor(reactive): pass service properties straight through
The service doesn't need to be passed to the effect because it can be
trivially accessed either by keeping a reference to the service as
returned by the createService call or by calling sig.use on the
service factory function.

Furthermore, TypeScript isn't able to infer the type of the service
when it's passed to the effect due to it being a circular type reference.
It is, however, able to infer the type of the service when it is
accessed using one of the two aforementioned methods.