fix(reactive): connect to a single, configured node as the uplink
Previously, the plan was to allow each wallet to connect to multiple
nodes as uplinks. However, there are numerous issues with this
architecture. For example, it would be difficult or impossible for the
wallet to have a public payment pointer. Outsourcing the provision of a
payment pointer to another entity requires complete trust.

By assuming that the owner of the wallet is the same person as the owner
of the node, we can greatly simplify the architecture for now.