refactor(reactive): run() returns actor instance instead of actor result
Previously, we would return the result of the first execution of the actor. One benefit of that approach was that ESLint would warn us if we weren't awaiting any promise returned by the actor.

But as actors have evolved from simple functions (originally called effects) to stateful actors, a lot of the instances where we were using actors as functions no longer make sense and should be replaced by plain async functions. It is therefore less important to capture the return values of actors, hence this API change.