feat(node): make routing table serializable
Previously, the routing table contained callback functions for sending packets to each given destination. In the case of destinations that were other Dassie nodes, this meant a separate set of callbacks for every node, even though we may never need to route a packet to that node.

With this change, the routing table just contains the node id for the next hop (which is just a string). This makes the routing table a much simpler data structure that can be examined easily. It should also take up less memory.

Overall, we are aiming for all long-lived state stores to always be serializable so this is another step towards that.