fix(logger): don't rely on absolute path for formatting stack traces
This code broke when updating Vite, because Vite added a feature where it tries to cleverly detect when you are referencing a module with the new URL("...", import.meta.url) syntax. But having the filesystem path baked in is strange anyways, so we just remove that and simplify the code. It could have a false positive if there is a packages/ folder somewhere else, but that seems like a minor problem since it's only used for color highlighting anyway.