refactor(node): make database a plain context value
The database reference isn't a piece of mutable state, therefore it should not be a signal.

We could make it an actor and give it its own lifecycle. But that would make not always available meaning we would have to handle the case where it isn't available yet and we are still waiting for it. This would make a lot of code in the application asynchronous for no good reason.

Instead, we make it a plain context value and have it inherit the reactor's lifecycle.