feat(dev): start nodes sequentially with some time in between
It was difficult to test the node bootstrapping process when all nodes appeared in the network simultaneously. When each node tried to map the network it found that nobody was peered with anybody yet. As a result, all nodes just peered with the bootstrap nodes.

With this delay, by the time the last node boots up, some nodes will already be peered. This allows most nodes to discover non-bootstrap nodes they can peer with.