feat(node): refer to endpoints in a consistent way
This commit takes the concept of "destination info" (i.e. everything you need to know to send an ILP packet somewhere) and further develops it into "endpoint info". This means we can use the same object on incoming packets and we can even reuse that information to send response packets simply by retaining the source endpoint information associated with the corresponding prepare packet.

We also move the routing table lookup to the routing feature which is where it properly belongs. For local ILP addresses like BTP, the routing operation is a simple lookup, i.e. a given address prefix corresponds to a specific endpoint. But for addresses corresponding to other Dassie nodes, the endpoint we should send to is a peer chosen from the set which is equally closest to the destination. Currently, we just pick the first one in the list but eventually we would apply a more sophisticated heuristic.