refactor(reactive): simplify reactive-trpc
This commit removes the old "remote reactive" system which used very complex and clunky types and relied on some internals and deprecated functionality of TRPC. The new version is much simpler and uses TRPC's intended public interface.

The node state tab from the dev environment has been removed. Most information shown on that tab is already viewable via the debug pages of the individual wallet UIs. Eventually, the wallet debug pages will be much more detailed than the dev interface was.

To the extent that we want to show node specific information directly in the dev interface, we should import the components from the wallet UI to avoid repeating that code.