feat(node): introduce config levels
This commit introduces the concept of "config levels", i.e. that the node may be at different stages of initialization. Some components may depend on certain configuration being present and will therefore not be loaded unless that configuration is available.

The main use case at the moment is to make the web UI available even though the node does not have a node key yet. This will allow the user to connect via the web UI to continue the setup process.