refactor(node): move as much config as possible to database
The exception here is the subnet configuration. Subnets (or 'settlement mechanisms' which may be a better term going forward) should be their own database table and not part of the configuration at all. The only reason they are currently part of the configuration is because I needed a way to instantiate subnets from the dev environment but nowadays the dev environment can just prepopulate the database accordingly.