fix(dev): remove warning about fsevents
This was a fun little bug so its story shall be told for those who wish to read it.

Once upon a time, a few commits ago, I switched the dev server from using the vite-node CLI to using vite-node programmatically. Since the switch the dev server would spit out an error while starting up but otherwise work fine without issues. The error read:

Failed to resolve import "fsevents" from "node_modules/.pnpm/rollup@3.25.3/node_modules/rollup/dist/es/shared/node-entry.js". Does the file exist?

Some investigation reveals that rollup tries to import fsevents. This fails on platforms other than Mac, but rollup catches the error and continues.

Looking at the vite source code, the error is printed during the module lookup which means that vite has no idea that this problem can be safely ignored and dutifully logs an error. But not always: In SSR mode, vite will not log the error. Instead, it leaves the error handling up to Node. Node throws an error, rollup catches it, and nothing is logged.

But why was vite-node running in SSR mode when I was using the CLI but not when I was using it programmatically?

As it turns out, vite-node will use the filename to decide whether to run something in SSR mode. If the file name has an extension (such as .ts), it will run in SSR mode. During the change to the dev server, I had inadvertently switched from referencing the dev server entry point with a file extension to referencing it without a file extension. And this explains the change in behavior.