feat(sqlite): remove serialize/deserialize
This feature would have made it very difficult to support more complex queries and provides little benefit. It abstracts away what the database actually returns which could make it harder to reason about code that interfaces with the database.

To the extent that data needs to be serialized/deserialized, the code that wraps the database (e.g. stores) is the right place to do so.