feat(reactive): better type for ApiProxy#ask
The previous version was always overriding the return type of the api method to make it asynchronous. Changing the return type of a function to be a Promise should be idempotent and should not modify functions that are already async. However, in practice it will break if the function signature is generic.

Instead, we can check if the function is already async and not touch the type if so. That preserves generics at least for async handlers.