feat(node): rework node discovery
The main functional change is that the current state of a node's node list is now represented by a hash, which means rather than downloading the full node list over and over, we can poll the hash and only download the list if the hash has changed.

This version also changes the node list to be just a list of node IDs without the full node state which is downloaded separately.

Finally, this commit refactors node discovery to be performed by several simple actors rather than one very complex one.