feat(dev): move restart logic out to start script
Previously, if there was a syntax error or some other exception stopping the dev server from starting up, the file change listener would not get registered and the dev server would not restart.

In this new version, the dev server will restart even after a previous run failed to parse or threw an exception.