fix(node): set up frontend middleware before starting https server
This was triggering a bug in the systemd socket activation. When the HTTPS server is closed, the socket will be closed as well, meaning that the next time it starts, it will fail.

The main bug still exists, this just avoids triggering it for now. The main bug will be fixed once we extract the environment-specific code (e.g. systemd integration) from the Dassie core.