feat(build): do not include install script on binary CDN
The install script is served from a different subdomain (sh.dassie.land)
than the compiled binaries (get.dassie.land). Once binary signing is
implemented, the latter will be much less sensitive than the former from
a security perspective.

Previously, we also served a copy of the install script on the binary
subdomain. However, since it is bad practice and possibly a security
concern to use that copy, it is better to remove it.