refactor(reactive): extract source revalidation from main actor loop
The goal of this change is to make it so we can have more control over
the source revalidation logic. This is done in order to be able to add
an abort feature whereby an actor boot process can be canceled either
due to a forced restart or because we already know that its tracked
reads are stale.