fix(oer): explicitly type string parsing method
This is the solution to a subtle TypeScript issue I was having.

The inferred type of various parse results in `app-dassie` was correctly
considered by `tsc` to be `string` but `typescript-eslint` inferred the
type as `error` which triggered
`@typescript-eslint/no-unsafe-assignment`. Eslint didn't report the
underlying type error because it is presumed that standalone `tsc` will
report the same ones but this wasn't the case here.

It turned out that the tsconfig for `app-dassie` was set to include
`@types/node` while the one for `lib-oer` did not include them. When
`tsc` compiled `app-dassie`, it used the project reference for `lib-oer`
and did not actually recompile `lib-oer` and thus it did not encounter
any problem. However, `typescript-eslint` apparently didn't use the
tsbuildinfo and thus descended into `lib-oer` in order to figure out the
return type of the parse function. Because TextEncoder/TextDecoder are
defined differently in `@types/node` than in `lib: "DOM"`, it
encountered an error while processing `lib-oer/src/string.ts`.

Explicitly typing the return value of the parse method for `OerString`
fixed the issue.