feat(incremental-check): make linting separately incremental
Previously, the linter would run based on which packages were processed
by TypeScript in the current run. This created several situations where
packages that hadn't been linted yet would not be. For instance, if a
`pnpm check` was canceled after TypeScript but before Eslint, none of
the packages that passed TypeScript would be linted.

This new implementation introduces an "Eslint tag" file whose modify
date corresponds to the last time Eslint ran successfully. It then
compares the date on that file to date the TypeScript cache file and
pnpm-lock.yaml. This ensures that if there have been any code or
dependency changes since the last successful Eslint run, a new check is
performed.