refactor(node): make migration `CREATE TABLE` statements reusable
It's quite useful to be able to refer to previous migrations' `CREATE
TABLE` statements in later migrations. (For down migrations.)

We previously just called the `up()` method of the previous migration to
accomplish but this only works if the the old migration creates a new
table, not if it actually migrates data from a previous version of the
table. This commit instead exports the `CREATE TABLE` statement as
string which is both easier to understand and more flexible.