feat(node): remove automatic registration
In the real network, registration as a tier-1 node is something that a
node admin would do explicitly.

Therefore, we remove the auto-registration functionality that was added
as a convenience during the early stages of development.

This commit also contains changes to pre-register nodes in the test
scenarios to ensure that they are still registered and able to connect.